+++ app/models/concerns/question_bank/bool_methods.rb
