+++ app/models/concerns/question_bank/choice_methods.rb
