+++ app/models/concerns/question_bank/essay_methods.rb
