+++ app/models/concerns/question_bank/multi_choice_methods.rb
