+++ app/models/concerns/question_bank/single_choice_methods.rb
