+++ app/models/concerns/question_bank/time_kind_scope.rb
